Major speed improvement in flat rendering

+++ src/r_sector.h
@@ -8,6 +8,6 @@
 #include "m_defs.h"
 
 // Draw a sector.
-void R_DrawSector(sector_t *sector, float left, float right);
+void R_DrawSector(sector_t *sector, uint16_t left, uint16_t right);
 
 #endif
